requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
distro==1.9.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.9.0
openai==1.69.0
pydantic==2.11.1
pydantic_core==2.33.0
python-dotenv==1.1.0
regex==2024.11.6
requests==2.32.3
sniffio==1.3.1
tavily-python==0.5.3
tiktoken==0.9.0
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.0
urllib3==2.3.0
pandas
google-api-python-client
genai
google